+++ bankfeeds/RELEASES.md
@@ -751,3 +751,13 @@ Based on:
 - Speakeasy CLI 1.69.1 (2.82.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v1.1.0] https://www.npmjs.com/package/@codat/bank-feeds/v/1.1.0 - bankfeeds
+
+## 2023-08-12 00:03:53
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.71.0 (2.83.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.1.1] bankfeeds
+### Releases
+- [NPM v1.1.1] https://www.npmjs.com/package/@codat/bank-feeds/v/1.1.1 - bankfeeds